+++ lib/src/fc_stateful.dart
@@ -1,9 +1,19 @@
 import 'package:flutter/widgets.dart';
+import 'package:full_context/src/fc_exception.dart';
 
 class FCStateful extends StatefulWidget {
-  const FCStateful({super.key, required this.builder, required this.onInit});
-  final Widget Function(BuildContext context) builder;
-  final void Function(BuildContext context) onInit;
+  const FCStateful({
+    super.key,
+    this.builder,
+    this.onInit,
+    this.afterInit,
+    this.child,
+  });
+
+  final Widget Function(BuildContext context)? builder;
+  final void Function(BuildContext context)? onInit;
+  final void Function(BuildContext context)? afterInit;
+  final Widget? child;
 
   @override
   State<FCStateful> createState() => _FCStatefulState();
@@ -13,9 +23,19 @@ class _FCStatefulState extends State<FCStateful> {
   @override
   void initState() {
     super.initState();
-    widget.onInit(context);
+    widget.onInit?.call(context);
+    WidgetsBinding.instance.addPostFrameCallback(
+      (_) => widget.afterInit?.call(context),
+    );
   }
 
   @override
-  Widget build(BuildContext context) => widget.builder(context);
+  Widget build(BuildContext context) {
+    final hasChild = widget.child != null;
+    if (hasChild == (widget.builder != null)) {
+      throw const FCException('Child or builder only');
+    }
+
+    return hasChild ? widget.child! : widget.builder!(context);
+  }
 }